% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_route.R
\name{plot_route}
\alias{plot_route}
\title{Plot a route map with connecting lines}
\usage{
plot_route(from, to, date, time, num = 5, route_index = 1)
}
\arguments{
\item{from}{Origin station ID.}

\item{to}{Destination station ID.}

\item{date}{Travel date in MM/DD/YYYY format.}

\item{time}{Travel time in HH:MM format.}

\item{num}{Number of routes to request.}

\item{route_index}{Index of the route to plot.}
}
\value{
A ggplot object.
}
\description{
Gets routes using \code{get_routes()} and plots one selected route by
connecting consecutive points with straight segments.
}
