% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_route_map.R
\name{plot_route_map}
\alias{plot_route_map}
\title{Plot a regional route map}
\usage{
plot_route_map(from, date, time, num = 3)
}
\arguments{
\item{from}{Origin station ID.}

\item{date}{Travel date in MM/DD/YYYY format.}

\item{time}{Travel time in HH:MM format.}

\item{num}{Number of routes requested per destination.}
}
\value{
A ggplot object.
}
\description{
Gets routes from the origin station to all stations in the region and
plots the first route for each destination.
}
