% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_data.R
\name{read_csv_data}
\alias{read_csv_data}
\title{Read saved CSV data}
\usage{
read_csv_data()
}
\value{
The saved dataset from ticino.alpine.transport.
}
\description{
Returns the saved dataset included in the ticino.alpine.transport package.
}
