% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_query_table.R
\name{route_query_table}
\alias{route_query_table}
\title{Route Query Table}
\usage{
route_query_table(
  query_date = "2025-06-20",
  query_times = c("07:30", "09:00", "12:00", "17:00", "19:00")
)
}
\arguments{
\item{query_date}{Selected query date.}

\item{query_times}{Selected query times.}
}
\value{
A tibble with route queries.
}
\description{
Route Query Table
}
