% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{saved_plot}
\alias{saved_plot}
\title{Ticino and Alpine transport stations}
\format{
A data frame with 15 rows and 11 variables:
\describe{
}
}
\source{
Prepared for the ticino.alpine.transport package.
}
\usage{
saved_plot
}
\description{
A dataset containing public transport stations in Ticino and the surrounding
Alpine region, including station identifiers, geographic coordinates, and
population data.
}
\keyword{datasets}
